+++ lib/puppet/rails/database/001_add_created_at_to_all_tables.rb
@@ -1,17 +1,13 @@
 class AddCreatedAtToAllTables < ActiveRecord::Migration
     def self.up
         ActiveRecord::Base.connection.tables.each do |t|
-            unless ActiveRecord::Base.connection.columns(t).collect {|c| c.name}.include?("created_at")
-                add_column t.to_s, :created_at, :datetime
-            end
+            add_column t.to_s, :created_at, :datetime unless ActiveRecord::Base.connection.columns(t).collect {|c| c.name}.include?("created_at")
         end
     end
 
     def self.down
         ActiveRecord::Base.connection.tables.each do |t|
-            unless ActiveRecord::Base.connection.columns(t).collect {|c| c.name}.include?("created_at")
-                remove_column t.to_s, :created_at
-            end
+            remove_column t.to_s, :created_at unless ActiveRecord::Base.connection.columns(t).collect {|c| c.name}.include?("created_at")
         end
     end
 end
